+++ app/src/main/java/com/fusetech/mobilleltarkotlin/ui/interfaces/LeltarListener.kt
@@ -1,5 +1,7 @@
 package com.fusetech.mobilleltarkotlin.ui.interfaces
 
+import java.math.BigDecimal
+
 interface LeltarListener {
     fun setRaktarText(code: String)
     fun setCikkText(code: String)
@@ -9,7 +11,7 @@ interface LeltarListener {
     fun raktarAdat(terulet: String)
     fun errorCode(code: String)
     fun clearAll()
-    fun mennyisegListener(quantity: Double)
+    fun mennyisegListener(quantity: BigDecimal)
     fun afterUpload()
     fun showData(listener: UpdateInterface, code: String)
     fun tabSwitch()
